[refactor] broker controller와 api controller 분리
