+++ src/main/java/kr/co/pawong/pwbe/chat/adapter/in/messaging/ChatMessageBrokerController.java
@@ -1,7 +1,7 @@
-package kr.co.pawong.pwbe.chat.adapter.in.api;
+package kr.co.pawong.pwbe.chat.adapter.in.messaging;
 
-import kr.co.pawong.pwbe.chat.adapter.in.api.dto.request.ChatMessageCreateRequest;
-import kr.co.pawong.pwbe.chat.adapter.in.api.dto.response.ChatMessageBrokerResponse;
+import kr.co.pawong.pwbe.chat.adapter.in.messaging.dto.request.ChatMessageCreateRequest;
+import kr.co.pawong.pwbe.chat.adapter.in.messaging.dto.response.ChatMessageBrokerResponse;
 import kr.co.pawong.pwbe.chat.application.port.in.CommandChatMessageDataUseCase;
 import kr.co.pawong.pwbe.chat.application.port.in.dto.ChatMessageDetail;
 import kr.co.pawong.pwbe.chat.enums.ChatMessageStatus;
@@ -9,17 +9,19 @@ import kr.co.pawong.pwbe.user.adapter.out.security.CustomUserDetails;
 import kr.co.pawong.pwbe.user.application.port.in.QueryNicknameUseCase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.simp.annotation.SendToUser;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PatchMapping;
 
 @Slf4j
 @Controller
 @RequiredArgsConstructor
-public class ChatMessageCommandController {
+public class ChatMessageBrokerController {
 
     private final CommandChatMessageDataUseCase commandChatMessageDataUseCase;
     private final QueryNicknameUseCase queryNicknameUseCase;
@@ -44,4 +46,5 @@ public class ChatMessageCommandController {
                         request.getCreatedAt()
                 ));
     }
+
 }
